ocr_service/requirements.txt
fastapi
uvicorn
pdf2image
python-multipart
paddleocr
paddlepaddle
scikit-learn
tqdm
pillow
numpy